#1 refactor: 코드 리팩토링을 통한 성능 개선
