+++ package.json
@@ -19,6 +19,7 @@
     "react-scripts": "5.0.1",
     "react-spinners": "0.13.8",
     "styled-components": "5.3.6",
+    "uuid": "^9.0.0",
     "web-vitals": "2.1.4"
   },
   "scripts": {
@@ -44,5 +45,8 @@
       "last 1 firefox version",
       "last 1 safari version"
     ]
+  },
+  "devDependencies": {
+    "@types/uuid": "^9.0.2"
   }
 }
