+++ src/components/common/Loading.tsx/Loading.tsx
@@ -1,5 +1,5 @@
-import React from "react";
 import { PulseLoader } from "react-spinners";
+
 import { color } from "../../../style/common/color";
 
 const Loading = () => {
